requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
contextlib2==21.6.0
datasets==1.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
entrypoints==0.4
execnet==1.9.0
fastjsonschema==2.16.1
filelock==3.7.1
fsspec==2022.7.1
huggingface-hub==0.0.8
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mock==4.0.3
multiprocess==0.70.13
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.6.6
notebook==6.4.12
numpy==1.21.6
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
path==16.4.0
path.py==12.5.0
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
py==1.11.0
pyahocorasick==1.4.4
pyarrow==9.0.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-shutil==1.7.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.7.25
requests==2.28.1
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.7.3
Send2Trash==1.8.0
seqeval==1.2.2
six==1.16.0
soupsieve==2.3.2.post1
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.10.3
tomli==2.0.1
torch==1.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.7.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
xxhash==3.0.0
zipp==3.8.1
